Respect architecture tiers for window globals
